bamlss Family for Zero-Inflated Discrete EGPD (ZIDEGPD)

Creates a family.bamlss object for fitting zero-inflated discrete Extended Generalized Pareto Distribution (ZIDEGPD) models with bamlss(). Uses dzidiscegpd, pzidiscegpd, qzidiscegpd, and rzidiscegpd internally.

Usage

zidegpd_bamlss(m = 1, ...)

Arguments

m

integer 1–4 selecting the G transformation type (see egpd_bamlss for details)

...

arguments passed to link specification

Value

An object of class family.bamlss

References

Ahmad, T. and Arshad, I. A. (2024). New flexible versions of extended generalized Pareto model for count data. arXiv preprint arXiv:2409.18719. https://arxiv.org/abs/2409.18719

Ahmad, T. and Hussain, A. (2025). Flexible model for varying levels of zeros and outliers in count data. arXiv preprint arXiv:2510.27365. https://arxiv.org/abs/2510.27365

See also